UI Overhaul and Functional Enhancements
Revamped the customer interface for a sleek, modern aesthetic that aligns with current design trends. Implemented a dark theme for improved readability and visual comfort. Initiated development of the 'Cart' class to streamline the shopping and checkout process. Enhanced data grid views to reflect the new, more contemporary style. This commit lays the groundwork for a more intuitive and visually appealing user experience.

+++ DAOS/CatalogDAO.cs
@@ -24,7 +24,7 @@ namespace EasyBizPos
                     {
                         ProductId = reader.GetInt32(0),
                         ProductName = reader.GetString(1),
-                        Price = reader.GetInt32(2),
+                        Price = reader.GetDecimal(2),
                         Stock = reader.GetInt32(3),
